# api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Schedule Adjuster API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/participants": {
      "get": {
        "tags": [
          "participant"
        ],
        "summary": "Get Participants",
        "operationId": "get_participants_participants_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ParticipantGet"
                  },
                  "type": "array",
                  "title": "Response Get Participants Participants Get"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "participant"
        ],
        "summary": "Update Participants",
        "operationId": "update_participants_participants_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/ParticipantCreate"
                },
                "type": "array",
                "title": "Participants Create"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ParticipantGet"
                  },
                  "type": "array",
                  "title": "Response Update Participants Participants Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "participant"
        ],
        "summary": "Add Participants",
        "operationId": "add_participants_participants_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/ParticipantCreate"
                },
                "type": "array",
                "title": "Participants Create"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ParticipantGet"
                  },
                  "type": "array",
                  "title": "Response Add Participants Participants Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/participant/{participant_id}": {
      "get": {
        "tags": [
          "participant"
        ],
        "summary": "Get Participant",
        "operationId": "get_participant_participant__participant_id__get",
        "parameters": [
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Participant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantGet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "participant"
        ],
        "summary": "Update Participant",
        "operationId": "update_participant_participant__participant_id__put",
        "parameters": [
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Participant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParticipantCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantGet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/participant": {
      "post": {
        "tags": [
          "participant"
        ],
        "summary": "Add Participant",
        "operationId": "add_participant_participant_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParticipantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantGet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/participant-profile/{participant_id}": {
      "get": {
        "tags": [
          "participant"
        ],
        "summary": "Get Participant Profile",
        "operationId": "get_participant_profile_participant_profile__participant_id__get",
        "parameters": [
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Participant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantProfileGet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "participant"
        ],
        "summary": "Update Participant Profile",
        "operationId": "update_participant_profile_participant_profile__participant_id__put",
        "parameters": [
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Participant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParticipantProfileCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantProfileGet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/participant-profile": {
      "post": {
        "tags": [
          "participant"
        ],
        "summary": "Add Participant Profile",
        "operationId": "add_participant_profile_participant_profile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParticipantProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantProfileGet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ParticipantAvailability": {
        "type": "string",
        "enum": [
          "YES",
          "MAYBE",
          "NO"
        ],
        "title": "ParticipantAvailability"
      },
      "ParticipantCreate": {
        "properties": {
          "schedule_id": {
            "type": "string",
            "title": "Schedule Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "toParticipate": {
            "type": "boolean",
            "title": "Toparticipate"
          },
          "availability": {
            "$ref": "#/components/schemas/ParticipantAvailability"
          }
        },
        "type": "object",
        "required": [
          "schedule_id",
          "user_id",
          "toParticipate"
        ],
        "title": "ParticipantCreate"
      },
      "ParticipantGet": {
        "properties": {
          "schedule_id": {
            "type": "string",
            "title": "Schedule Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "toParticipate": {
            "type": "boolean",
            "title": "Toparticipate"
          },
          "availability": {
            "$ref": "#/components/schemas/ParticipantAvailability"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user": {
            "$ref": "#/components/schemas/UserGet"
          },
          "prodile": {
            "$ref": "#/components/schemas/ParticipantProfileGet"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "schedule_id",
          "user_id",
          "toParticipate",
          "id",
          "user",
          "prodile",
          "created_at"
        ],
        "title": "ParticipantGet"
      },
      "ParticipantProfileCreate": {
        "properties": {
          "participant_id": {
            "type": "string",
            "title": "Participant Id"
          },
          "intProp1": {
            "title": "Intprop1",
            "type": "integer"
          },
          "intProp2": {
            "title": "Intprop2",
            "type": "integer"
          },
          "intProp3": {
            "title": "Intprop3",
            "type": "integer"
          },
          "listProp1": {
            "title": "Listprop1",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "listProp2": {
            "title": "Listprop2",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "listProp3": {
            "title": "Listprop3",
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object",
        "required": [
          "participant_id"
        ],
        "title": "ParticipantProfileCreate"
      },
      "ParticipantProfileGet": {
        "properties": {
          "participant_id": {
            "type": "string",
            "title": "Participant Id"
          },
          "intProp1": {
            "title": "Intprop1",
            "type": "integer"
          },
          "intProp2": {
            "title": "Intprop2",
            "type": "integer"
          },
          "intProp3": {
            "title": "Intprop3",
            "type": "integer"
          },
          "listProp1": {
            "title": "Listprop1",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "listProp2": {
            "title": "Listprop2",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "listProp3": {
            "title": "Listprop3",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "participant_id",
          "created_at"
        ],
        "title": "ParticipantProfileGet"
      },
...
      "UserGet": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "birthday": {
            "title": "Birthday",
            "type": "string",
            "format": "date"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "role",
          "name",
          "email",
          "id",
          "created_at"
        ],
        "title": "UserGet"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "USER",
          "ADMIN"
        ],
        "title": "UserRole"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}